+++ docs/developers/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 📅 Revision history for HelPA
 
 ## 0.4.6.2 -- 2025-01-12
+* Add examples
 * Add `SpecHook` for parallel tests
 
 ## 0.4.6.1 -- 2025-01-11