+++ docs/developers/CHANGELOG.md
@@ -1,7 +1,12 @@
 # 📅 Revision history for HelPA
 
 ## 0.4.6.2 -- 2025-01-12
-* Add examples
+* Add `te` examples
+* Add `bbj` examples
+* Add `fbf` examples
+* Add `sico` examples
+* Add `sblasm` examples
+* Add `hapyli` examples
 * Add `SpecHook` for parallel tests
 
 ## 0.4.6.1 -- 2025-01-11